requirements.txt
fastapi==0.111.0
pydantic==2.8.2
aiohttp==3.9.3
PyJWT==2.6.0
aiofiles==22.1.0
redis==4.6.0
SQLAlchemy==2.0.30
uvicorn==0.20.0
python-multipart==0.0.7
httpx==0.24.0
loguru==0.7.0
captcha==0.5.0
pydantic-settings==2.3.3
bcrypt==4.1.3
aiomysql==0.2.0

aiosqlite==0.19.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
asgi-lifespan==2.1.0